This is currently only used for linalg.generic ops and is a means to
vectorize these operations by default.