@hanhanW 2021-12-15
This depends on https://reviews.llvm.org/D115820

Performance on local Pixel 4.

| Model            | Before  | After   |
| ---------------- | ------- | ------- |
| DeepLabV3        | 964 ms  | 954 ms  |
| MobileBertSquad  | 682 ms  | 682 ms  |
| MobileNetV2      | 177 ms  | 176 ms  |
| MobileNetV3Small | 47.6 ms | 47.5 ms |
| MobileSSD        | 365 ms  | 368 ms  |
| PoseNet          | 884 ms  | 884 ms  |

There is a slightly improvement (10ms) in DeepLabV3.

After running many times, I think the regression in MobileSSD is true (~2 ms). I think it's fine for now.